Get point data from a c3d object

Get the point data of an c3d object in a data frame.

Usage

c3d_data(x, format = "wide")

Arguments

x

A c3d object, as imported by c3d_read.

format

Either "wide" (default), "long", or "longest" to determine the format of the resulting data frame. The wide format has Long format has three columns per point (x,y,z), the long format has three rows per frame per point (x,y,z). The longest format has one row per data entry.

Value

A data frame of class c3d_data with the c3d point data. The structure of the data frame depends on the 'format' argument.

Details

The point data of imported c3d objects in c3dr is saved as a list of lists. This is good for internal handling, but for analysis a table format (a data frame) is often more convenient.c3d_data() returns the point data from an imported c3d object as a data frame.

Analyses of data frames may require them to have different formats. For c3d_data output, different data formats ("wide", "long", "longest") are available. See the format argument for more details. You can convert between different formats with c3d_convert.

Examples

# Import example data
d <- c3d_read(c3d_example())

# wide format (default)
w <- c3d_data(d)
head(w)
#>     L_IAS_x  L_IAS_y  L_IAS_z   L_IPS_x  L_IPS_y  L_IPS_z   R_IPS_x  R_IPS_y
#> 1 -220.1226 306.4248 846.3361 -398.1731 237.0688 872.8574 -392.8751 146.2103
#> 2 -212.4696 306.5356 844.6985 -390.7831 237.8691 871.2827 -385.6659 147.1048
#> 3 -204.8696 306.6555 843.2342 -383.1857 238.6758 869.6015 -378.4103 148.0202
#> 4 -197.1952 306.8035 841.6127 -375.7068 239.5434 868.1396 -371.1021 148.9276
#> 5 -189.6655 307.0628 840.1692 -368.1680 240.4141 866.6924 -363.7109 149.8643
#> 6 -182.0234 307.3520 838.7117 -360.4463 241.2611 865.1455 -356.2582 150.8374
#>    R_IPS_z   R_IAS_x  R_IAS_y  R_IAS_z     SNJ_x    SNJ_y    SNJ_z     SXS_x
#> 1 880.3161 -220.9995 64.67303 844.9208 -216.6414 201.7562 1269.208 -184.2733
#> 2 878.9836 -213.5632 64.95774 843.6115 -209.2227 202.5531 1267.821 -176.9668
#> 3 877.7510 -206.1010 65.27660 842.5367 -201.6737 203.3655 1266.284 -169.3281
#> 4 876.4678 -198.4038 65.66290 841.2524 -194.1267 204.1923 1264.927 -162.0173
#> 5 875.2322 -190.7461 66.13321 840.3125 -186.5323 204.9816 1263.396 -154.3956
#> 6 874.0587 -183.0446 66.55704 839.0317 -179.0034 205.8802 1262.149 -146.9180
#>      SXS_y    SXS_z     TV8_x    TV8_y    TV8_z     CV7_x    CV7_y    CV7_z
#> 1 205.9899 1119.616 -411.1556 181.7056 1137.173 -348.7630 190.4895 1354.547
#> 2 206.5667 1118.196 -403.7221 182.3904 1135.906 -341.2098 191.3307 1353.139
#> 3 207.1396 1116.586 -396.2882 183.1377 1134.625 -333.6015 192.1670 1351.741
#> 4 207.7761 1115.223 -388.7892 183.8530 1133.361 -326.0073 193.0315 1350.351
#> 5 208.4521 1113.678 -381.2719 184.6602 1132.119 -318.3973 193.9117 1348.993
#> 6 209.1185 1112.445 -373.6935 185.4699 1130.915 -310.8057 194.8010 1347.726
#>    R_SCAP_x R_SCAP_y R_SCAP_z   L_HDF_x  L_HDF_y  L_HDF_z   L_HDB_x  L_HDB_y
#> 1 -395.8241 60.79993 1196.906 -165.0246 230.6246 1441.551 -234.5870 269.4478
#> 2 -388.4307 61.51912 1195.727 -157.5963 231.4405 1439.996 -227.1441 270.2314
#> 3 -380.9545 62.29219 1194.465 -150.1807 232.2044 1438.783 -219.6844 271.0766
#> 4 -373.4888 63.04057 1193.297 -142.6875 233.0275 1437.414 -212.1976 271.9120
#> 5 -365.9877 63.86989 1192.030 -135.2145 233.9135 1436.036 -204.6188 272.9810
#> 6 -358.4762 64.69230 1190.908 -127.7784 234.7733 1435.001 -197.2276 273.7422
#>    L_HDB_z   R_HDB_x  R_HDB_y  R_HDB_z   R_HDF_x  R_HDF_y  R_HDF_z   L_FTC_x
#> 1 1455.001 -226.5189 119.8213 1451.025 -159.8333 168.3401 1439.981 -263.7520
#> 2 1453.467 -219.2512 120.7211 1449.627 -152.5004 169.1234 1438.389 -256.7881
#> 3 1452.255 -211.9548 121.5144 1448.201 -145.0881 169.9457 1437.063 -249.6585
#> 4 1450.802 -204.6025 122.3854 1446.849 -137.8438 170.6613 1435.801 -242.3391
#> 5 1449.382 -197.2597 123.3404 1445.504 -130.1578 171.5359 1434.287 -234.6830
#> 6 1448.229 -189.8334 124.2127 1444.327 -122.9863 172.4056 1433.200 -226.8020
#>    L_FTC_y  L_FTC_z L_WAND1_x L_WAND1_y L_WAND1_z    L_FLE_x  L_FLE_y  L_FLE_z
#> 1 367.2804 752.2808 -180.9762  371.5085  554.2516 -133.92155 350.0119 449.4388
#> 2 367.6512 750.3760 -175.5038  372.8508  552.0001 -127.53288 350.2077 447.3870
#> 3 368.1453 748.5175 -169.6553  374.2023  549.6017 -120.83626 350.4556 445.5355
#> 4 368.6311 746.8800 -163.2146  375.5533  547.8017 -113.62605 350.4675 444.1317
#> 5 369.1276 745.3724 -155.4749  376.9098  546.1851 -105.59972 350.6089 443.1502
#> 6 369.8208 744.0350 -146.5480  377.8416  544.9144  -96.63529 351.0294 442.5793
#>      L_FME_x  L_FME_y  L_FME_z    L_FAX_x  L_FAX_y  L_FAX_z   L_TTC_x  L_TTC_y
#> 1 -107.18948 238.5570 435.9054 -121.99879 356.5338 381.3046 -59.09079 314.1751
#> 2 -100.65621 237.9747 434.0668 -113.22730 356.5356 380.7096 -50.35317 314.3902
#> 3  -93.69539 237.4486 432.1809 -104.51640 356.5098 379.9195 -41.71322 314.5672
#> 4  -86.27704 236.8322 430.6884  -95.95802 356.4550 378.9916 -33.11832 314.7371
#> 5  -78.73642 236.2442 429.9315  -87.61524 356.4284 377.9061 -24.70577 314.7635
#> 6  -70.88741 235.7930 429.3256  -78.81496 356.8516 376.2351 -16.19576 314.7452
#>    L_TTC_z L_WAND2_x L_WAND2_y L_WAND2_z    L_FAL_x  L_FAL_y  L_FAL_z  L_TAM_x
#> 1 382.2960 -96.04347  368.4831  274.0905 -10.577349 335.3952 76.46465 10.41641
#> 2 382.6820 -84.57689  368.8843  273.8767   5.028977 335.6764 77.85229 25.05615
#> 3 383.0185 -73.57381  369.2250  273.5958  19.708595 335.9773 79.20754 38.90612
#> 4 383.1797 -63.22645  369.2780  273.1905  33.529842 336.2560 80.40594 51.97777
#> 5 383.3443 -53.63693  369.0824  272.6349  46.583862 336.4469 81.40598 64.18738
#> 6 383.2223 -44.52882  368.7002  271.8571  58.274323 336.4017 81.80460 75.51200
#>    L_TAM_y  L_TAM_z      L_FCC_x  L_FCC_y  L_FCC_z  L_FM1_x  L_FM1_y  L_FM1_z
#> 1 258.7448 108.8423 -32.02432632 282.1060 23.57637 135.3946 291.3752 106.1674
#> 2 258.9918 110.8000 -15.45973301 282.4953 24.37029 150.3147 290.3775 110.7726
#> 3 259.3221 112.4818   0.06008298 282.8171 25.38456 164.2544 289.3691 115.1992
#> 4 259.6789 113.9113  14.53176975 283.1865 26.02039 176.9873 288.2792 118.8425
#> 5 259.9023 115.0040  28.02707672 283.3470 26.42657 189.2864 287.5384 121.4435
#> 6 259.8813 115.7052  40.24918747 283.2290 26.72655 200.4272 286.9825 123.6483
#>    L_FM5_x  L_FM5_y  L_FM5_z   R_FTC_x  R_FTC_y  R_FTC_z R_WAND1_x R_WAND1_y
#> 1  97.8722 368.0651 77.41405 -297.6747 11.44923 754.3581 -316.3554  19.85999
#> 2 113.9963 367.4883 80.86826 -291.3019 11.67470 753.3961 -310.8152  19.36007
#> 3 129.2370 366.8577 83.89771 -284.7861 11.88232 752.4771 -305.0969  19.05180
#> 4 143.5583 366.4146 86.45792 -278.0583 12.08494 751.6553 -299.2249  19.22254
#> 5 156.7796 365.9593 88.40491 -271.1957 12.43242 750.8671 -293.1631  19.51930
#> 6 168.7515 365.5351 89.85886 -264.2056 12.79618 750.0839 -286.9101  19.82918
#>   R_WAND1_z   R_FLE_x  R_FLE_y  R_FLE_z   R_FME_x  R_FME_y  R_FME_z   R_FAX_x
#> 1  627.7543 -392.5053 62.28204 432.2710 -386.3387 168.2844 420.1111 -441.1790
#> 2  626.9677 -388.2951 62.21341 431.9850 -382.0948 168.2361 419.7383 -437.4213
#> 3  626.1819 -383.8289 62.08937 431.7456 -377.6484 168.1340 419.2292 -433.3869
#> 4  625.6271 -379.1577 61.93435 431.4996 -372.8698 167.9975 418.8507 -429.2064
#> 5  625.1700 -373.7126 61.43665 431.6674 -368.0770 167.9380 418.3861 -424.7871
#> 6  624.7526 -368.6255 61.26788 431.3504 -362.9131 167.7041 417.8956 -420.2483
#>    R_FAX_y  R_FAX_z   R_TTC_x  R_TTC_y  R_TTC_z R_WAND2_x R_WAND2_y R_WAND2_z
#> 1 58.43905 376.7283 -382.6898 104.5305 351.7569 -513.4461  84.31416  196.9792
#> 2 58.31371 376.9108 -379.2307 104.3874 351.2991 -511.5775  84.02758  197.9374
#> 3 58.16567 377.1750 -375.5656 104.2098 350.8492 -509.5206  83.72875  199.0019
#> 4 58.00101 377.3995 -371.7306 104.1192 350.3814 -507.3555  83.41702  200.0947
#> 5 57.76202 377.6834 -367.7058 103.9782 349.9180 -505.0082  82.98409  201.2887
#> 6 57.55413 377.9538 -363.4827 103.8096 349.4844 -502.5269  82.57461  202.4673
#>     R_FAL_x  R_FAL_y  R_FAL_z   R_TAM_x  R_TAM_y  R_TAM_z   R_FCC_x  R_FCC_y
#> 1 -554.6070 113.3102 78.51108 -517.1957 192.9693 89.40974 -605.9172 167.0107
#> 2 -553.9764 112.9010 79.85748 -517.4850 193.0359 89.42204 -605.6677 166.4978
#> 3 -553.3301 112.3712 81.15167 -516.2042 192.2898 90.71317 -605.3972 165.9186
#> 4 -553.3693 111.6689 81.76969 -514.7889 191.4312 92.07403 -605.0107 165.2526
#> 5 -552.5394 111.0167 83.40919 -513.2784 190.5146 93.47956 -604.6310 164.5759
#> 6 -551.5978 110.3237 85.19250 -511.5644 189.3739 95.01070 -604.1198 163.8309
#>    R_FCC_z   R_FM1_x  R_FM1_y  R_FM1_z   R_FM5_x  R_FM5_y  R_FM5_z   L_HM5_x
#> 1 50.45440 -415.4895 162.1822 24.36582 -467.8351 71.08191 19.22491 -457.8361
#> 2 52.35530 -415.9909 162.0967 23.94669 -467.7528 71.03170 19.21836 -453.8291
#> 3 54.35469 -416.0649 162.0302 23.92010 -467.6797 70.97117 19.20053 -449.6225
#> 4 56.58235 -416.1300 161.9581 23.91689 -469.0358 71.40529 17.95319 -445.0983
#> 5 58.98922 -416.2096 161.8565 23.92274 -469.0214 71.34734 17.90706 -440.2917
#> 6 61.38317 -416.2549 161.7771 23.94778 -468.9145 71.28360 17.94883 -434.2212
#>    L_HM5_y  L_HM5_z   L_HM2_x  L_HM2_y  L_HM2_z   L_UHE_x  L_UHE_y  L_UHE_z
#> 1 495.2370 728.3018 -377.8411 453.7730 688.1711 -420.6162 495.7649 801.8030
#> 2 495.7131 727.7291 -374.0018 454.8627 686.7296 -415.9475 496.3765 800.8562
#> 3 496.3557 727.2435 -369.9855 455.8685 685.3533 -411.0641 496.9818 799.9400
#> 4 496.8854 726.7498 -365.6416 456.9299 684.0995 -405.9654 497.6008 799.0549
#> 5 497.4112 726.0722 -360.9733 457.9977 682.9140 -400.6318 498.2273 798.1677
#> 6 498.6086 725.9686 -356.1921 458.8551 681.7466 -395.0639 498.8650 797.3062
#>     L_RSP_x  L_RSP_y  L_RSP_z L_WAND4_x L_WAND4_y L_WAND4_z   L_HLE_x  L_HLE_y
#> 1 -371.7066 441.3212 784.5175 -378.1795  479.0819  927.7591 -386.4662 448.4879
#> 2 -366.8741 442.2759 783.0983 -372.3811  479.5789  926.4894 -379.9829 448.9615
#> 3 -361.7965 443.3029 781.6329 -366.6931  480.3084  925.2871 -373.4902 449.3326
#> 4 -356.6414 444.1617 780.3854 -360.5961  480.8386  924.0024 -366.9463 449.9247
#> 5 -351.1692 444.9383 779.0714 -354.4337  481.5292  922.7555 -360.0651 450.3554
#> 6 -345.5571 445.8037 777.7964 -348.0975  482.2153  921.6100 -353.2113 450.8650
#>    L_HLE_z   L_HME_x  L_HME_y  L_HME_z L_WAND3_x L_WAND3_y L_WAND3_z    R_HM5_x
#> 1 1017.401 -407.9913 368.9505 996.1462 -360.1560  433.5987  1108.638 -52.563515
#> 2 1015.954 -401.2656 369.3148 994.7280 -353.3919  434.1775  1107.125 -41.163475
#> 3 1014.667 -394.4433 369.6541 993.4374 -346.5003  434.7382  1105.699 -29.981133
#> 4 1013.497 -386.7114 369.2739 991.4387 -339.4201  435.3300  1104.258 -19.080381
#> 5 1012.178 -379.5547 369.6029 990.1309 -332.4544  435.9444  1102.935  -8.331214
#> 6 1010.916 -372.2626 369.9787 988.8334 -325.1070  436.6194  1101.575   2.045109
#>      R_HM5_y  R_HM5_z  R_HM2_x   R_HM2_y  R_HM2_z   R_UHE_x   R_UHE_y  R_UHE_z
#> 1 -105.65646 767.4525 18.09681 -41.25887 797.3684 -82.49286 -106.4504 839.3554
#> 2 -104.28706 768.5121 29.04011 -40.11528 799.9297 -72.18317 -105.4229 839.6418
#> 3 -102.88853 769.5044 39.77573 -39.02583 802.3937 -62.13435 -104.4321 840.3021
#> 4 -101.48241 770.6132 50.32013 -37.97644 804.8308 -52.56231 -103.6892 841.1437
#> 5 -100.08453 771.6097 60.93653 -36.61471 807.4357 -42.82722 -102.6136 841.4954
#> 6  -98.65206 772.5919 70.96821 -35.45071 809.6336 -33.49717 -101.4092 842.1796
#>     R_RSP_x   R_RSP_y  R_RSP_z  R_WAND4_x R_WAND4_y R_WAND4_z   R_HLE_x
#> 1 -54.86601 -34.87722 851.6902 -119.91284 -98.70297  910.0037 -224.3468
#> 2 -44.64488 -33.90419 853.0445 -110.45589 -97.89973  910.2659 -216.2131
#> 3 -34.64185 -32.87292 854.2796 -101.27998 -97.07436  910.3013 -208.0463
#> 4 -24.92161 -31.77574 855.5604  -92.06359 -96.15163  910.3701 -200.0863
#> 5 -15.32224 -30.68514 856.7554  -83.25047 -95.20078  910.6467 -192.1246
#> 6  -5.87497 -29.61648 857.9356  -76.50838 -95.18553  909.9816 -184.2547
#>     R_HLE_y  R_HLE_z   R_HME_x   R_HME_y  R_HME_z R_WAND3_x R_WAND3_y R_WAND3_z
#> 1 -86.36720 1016.870 -259.3438 -17.13619 977.2628 -233.6956 -51.29948  1130.602
#> 2 -85.94968 1015.901 -250.6690 -16.66577 975.9108 -225.9235 -50.57361  1129.593
#> 3 -85.43781 1014.913 -242.2320 -16.10110 974.6073 -218.1034 -49.82697  1128.553
#> 4 -84.83279 1014.024 -233.8171 -15.41979 973.3947 -210.3974 -48.95750  1127.515
#> 5 -84.20478 1013.156 -225.5507 -14.71915 972.3180 -202.8000 -48.07650  1126.723
#> 6 -83.45292 1012.267 -217.2875 -13.87957 971.1982 -194.9735 -47.12794  1125.655
#>     L_SAJ_x  L_SAJ_y  L_SAJ_z   R_SAJ_x  R_SAJ_y  R_SAJ_z
#> 1 -308.2423 369.5854 1294.401 -255.6461 18.73160 1295.157
#> 2 -300.8522 370.3304 1292.805 -248.1305 19.66945 1293.777
#> 3 -293.5054 371.1862 1291.279 -240.6772 20.63699 1292.478
#> 4 -286.0263 372.0377 1289.798 -233.1048 21.55496 1291.207
#> 5 -278.4725 372.9155 1288.340 -225.4429 22.43432 1289.966
#> 6 -270.9297 373.8162 1287.003 -217.8655 23.34610 1288.820

# long format
l <- c3d_data(d, format = "long")
head(l)
#>   frame type     L_IAS     L_IPS     R_IPS      R_IAS       SNJ       SXS
#> 1     1    x -220.1226 -398.1731 -392.8751 -220.99945 -216.6414 -184.2733
#> 2     1    y  306.4248  237.0688  146.2103   64.67303  201.7562  205.9899
#> 3     1    z  846.3361  872.8574  880.3161  844.92078 1269.2075 1119.6162
#> 4     2    x -212.4696 -390.7831 -385.6659 -213.56319 -209.2227 -176.9668
#> 5     2    y  306.5356  237.8691  147.1048   64.95774  202.5531  206.5667
#> 6     2    z  844.6985  871.2827  878.9836  843.61145 1267.8208 1118.1957
#>         TV8       CV7     R_SCAP     L_HDF     L_HDB     R_HDB     R_HDF
#> 1 -411.1556 -348.7630 -395.82407 -165.0246 -234.5870 -226.5189 -159.8333
#> 2  181.7056  190.4895   60.79993  230.6246  269.4478  119.8213  168.3401
#> 3 1137.1731 1354.5465 1196.90637 1441.5514 1455.0013 1451.0249 1439.9808
#> 4 -403.7221 -341.2098 -388.43066 -157.5963 -227.1441 -219.2512 -152.5004
#> 5  182.3904  191.3307   61.51912  231.4405  270.2314  120.7211  169.1234
#> 6 1135.9058 1353.1388 1195.72693 1439.9963 1453.4669 1449.6267 1438.3892
#>       L_FTC   L_WAND1     L_FLE     L_FME     L_FAX     L_TTC   L_WAND2
#> 1 -263.7520 -180.9762 -133.9216 -107.1895 -121.9988 -59.09079 -96.04347
#> 2  367.2804  371.5085  350.0119  238.5570  356.5338 314.17508 368.48309
#> 3  752.2808  554.2516  449.4388  435.9054  381.3046 382.29599 274.09045
#> 4 -256.7881 -175.5038 -127.5329 -100.6562 -113.2273 -50.35317 -84.57689
#> 5  367.6512  372.8508  350.2077  237.9747  356.5356 314.39023 368.88431
#> 6  750.3760  552.0001  447.3870  434.0668  380.7096 382.68198 273.87668
#>        L_FAL     L_TAM     L_FCC    L_FM1     L_FM5      R_FTC    R_WAND1
#> 1 -10.577349  10.41641 -32.02433 135.3946  97.87220 -297.67468 -316.35541
#> 2 335.395172 258.74478 282.10599 291.3752 368.06506   11.44923   19.85999
#> 3  76.464653 108.84234  23.57637 106.1674  77.41405  754.35809  627.75427
#> 4   5.028977  25.05615 -15.45973 150.3147 113.99632 -291.30191 -310.81516
#> 5 335.676392 258.99179 282.49527 290.3775 367.48834   11.67470   19.36007
#> 6  77.852287 110.79998  24.37029 110.7726  80.86826  753.39612  626.96771
#>        R_FLE     R_FME      R_FAX     R_TTC    R_WAND2      R_FAL      R_TAM
#> 1 -392.50528 -386.3387 -441.17905 -382.6898 -513.44611 -554.60699 -517.19568
#> 2   62.28204  168.2844   58.43905  104.5305   84.31416  113.31016  192.96931
#> 3  432.27097  420.1111  376.72830  351.7569  196.97916   78.51108   89.40974
#> 4 -388.29510 -382.0948 -437.42130 -379.2307 -511.57745 -553.97638 -517.48499
#> 5   62.21341  168.2361   58.31371  104.3874   84.02758  112.90102  193.03587
#> 6  431.98502  419.7383  376.91083  351.2991  197.93744   79.85748   89.42204
#>       R_FCC      R_FM1      R_FM5     L_HM5     L_HM2     L_UHE     L_RSP
#> 1 -605.9172 -415.48947 -467.83508 -457.8361 -377.8411 -420.6162 -371.7066
#> 2  167.0107  162.18219   71.08191  495.2370  453.7730  495.7649  441.3212
#> 3   50.4544   24.36582   19.22491  728.3018  688.1711  801.8030  784.5175
#> 4 -605.6677 -415.99088 -467.75284 -453.8291 -374.0018 -415.9475 -366.8741
#> 5  166.4978  162.09671   71.03170  495.7131  454.8627  496.3765  442.2759
#> 6   52.3553   23.94669   19.21836  727.7291  686.7296  800.8562  783.0983
#>     L_WAND4     L_HLE     L_HME   L_WAND3      R_HM5     R_HM2      R_UHE
#> 1 -378.1795 -386.4662 -407.9913 -360.1560  -52.56351  18.09681  -82.49286
#> 2  479.0819  448.4879  368.9505  433.5987 -105.65646 -41.25887 -106.45036
#> 3  927.7591 1017.4007  996.1462 1108.6378  767.45245 797.36841  839.35541
#> 4 -372.3811 -379.9829 -401.2656 -353.3919  -41.16348  29.04011  -72.18317
#> 5  479.5789  448.9615  369.3148  434.1775 -104.28706 -40.11528 -105.42285
#> 6  926.4894 1015.9540  994.7280 1107.1255  768.51208 799.92969  839.64185
#>       R_RSP    R_WAND4      R_HLE      R_HME    R_WAND3     L_SAJ      R_SAJ
#> 1 -54.86601 -119.91284 -224.34680 -259.34378 -233.69557 -308.2423 -255.64606
#> 2 -34.87722  -98.70297  -86.36720  -17.13619  -51.29948  369.5854   18.73160
#> 3 851.69025  910.00372 1016.87018  977.26276 1130.60181 1294.4005 1295.15723
#> 4 -44.64488 -110.45589 -216.21309 -250.66902 -225.92348 -300.8522 -248.13048
#> 5 -33.90419  -97.89973  -85.94968  -16.66577  -50.57361  370.3304   19.66945
#> 6 853.04449  910.26593 1015.90094  975.91083 1129.59290 1292.8052 1293.77747

# longest format
ll <- c3d_data(d, format = "longest")
head(ll)
#>   frame type point     value
#> 1     1    x L_IAS -220.1226
#> 2     1    y L_IAS  306.4248
#> 3     1    z L_IAS  846.3361
#> 4     1    x L_IPS -398.1731
#> 5     1    y L_IPS  237.0688
#> 6     1    z L_IPS  872.8574